Drawing state extracted.
