+++ src/sprite/mod.rs
@@ -2,8 +2,10 @@ mod base;
 mod behavior;
 mod model;
 mod outline;
+mod drawing_state;
 
-pub use base::{DrawingState, Sprite};
+pub use base::Sprite;
+pub use drawing_state::DrawingState;
 pub use behavior::{BehaviorManager, Hover};
 pub use model::SpriteMutation;
 pub use outline::Outline;
